requirements.txt
dnspython
Flask
gunicorn
